First attempt, working
This works, but we likely can revert to the regular ol commands with
`node_modules/.bin` on the path, rather than referring to binaries with
the `node_modules/.bin/` prefix.

This was originally not working because the two oldest versions of
`polyfill-library` have postinstall scripts, which fail to run in Google
AppEngine Standard. This commit removes those libraries, which is a
horrible fix, but we don't have a ton of choice right now.